+++ docs/guides/changelog.md
@@ -17,6 +17,7 @@ For migration guidance and non-schema changes, see the [project changelog](https
 - Added attribute `chalk_cluster_gateway.certificate_issuer_ref.group` (`string`).
 - Added attribute `chalk_cluster_gateway.certificate_issuer_ref.kind` (`string`).
 - Added attribute `chalk_cluster_gateway.certificate_issuer_ref.name` (`string`).
+- Added attribute `chalk_telemetry.runtime` (`string`).
 - Added attribute `chalk_unmanaged_environment.dataplane_db_secret` (`string`).
 
 ## v1.0.5